[로깅 추출] : LoggingModel을 엑셀에 작성하는 기능 구현

+++ excel/src/main/java/me/sun/excel/RowWrapper.java
@@ -35,7 +35,7 @@ public class RowWrapper {
         return rowWrapper;
     }
 
-    public RowWrapper writeCell(Long value) {
+    public RowWrapper writeCell(Integer value) {
         Cell cell = getCell();
         cell.setCellValue(value);
         applyStyle(cell);
